requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
distlib==0.3.6
Django==4.1.6
django-cors-headers==3.13.0
django-rest-framework==0.1.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
filelock==3.9.0
heroku==0.1.4
idna==3.4
mysqlclient==2.1.1
Pillow==9.4.0
platformdirs==2.6.2
psycopg2==2.9.5
python-dateutil==1.5
pytz==2022.7.1
requests==2.28.2
SpeechRecognition==3.9.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.18.0
